Renaming types.
